GradioTranscription/3 Components GradioTranscription/requirements.txt
gradio>=4.0.0
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0
pydub>=0.25.1
ffmpeg-python>=0.2.0
numpy>=1.20.0
tqdm>=4.65.0
more-itertools>=9.1.0